Добавлены и реализованы интерфейсы. Рефакторинг.
